+++ LearnXNAGame/LearnXNAGame/Maps/IMapCell.cs
@@ -5,8 +5,8 @@ namespace LearnXNAGame.Maps
 	public interface IMapCell
 	{
 		IMapCellNeighbor<IMapCell> Neighbors { get; }
-		float X { get; set; }
-		float Y { get; set; }
+		int X { get; set; }
+		int Y { get; set; }
 		void Draw(SpriteBatch sb, float offsetX = 0f, float offsetY = 0f);
 
 		bool IsHover { get; set; }
